合并wf

+++ src/main/java/com/xxxx/supermarket/entity/Goods.java
@@ -18,11 +18,11 @@ public class Goods {
 
     private String producer;//供应商
 
-    private Float purchasingPrice;//采购价格
+    private Double purchasingPrice;//采购价格
 
     private String remarks;//备注
 
-    private Float sellingPrice;//出售价格
+    private Double sellingPrice;//出售价格
 
     private String unit;//商品单位
